requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
bitarray==1.1.0
boto3==1.9.231
botocore==1.12.231
certifi==2019.11.28
chardet==3.0.4
Click==7.0
configparser==4.0.2
contextlib2==0.5.5
docutils==0.15.2
Flask==1.1.1
funcsigs==1.0.2
futures==3.3.0
gevent==1.5a2
geventhttpclient-wheels==1.3.1.dev2
greenlet==0.4.15
idna==2.8
importlib-metadata==0.22
impyla==0.16.2
itsdangerous==1.1.0
JayDeBeApi==1.1.1
Jinja2==2.10.3
jmespath==0.9.4
JPype1==0.6.3
locustio==0.13.5
MarkupSafe==1.1.1
more-itertools==5.0.0
msgpack-python==0.5.6
numpy==1.16.5
packaging==19.1
pathlib2==2.3.4
pluggy==0.13.0
py==1.8.0
pyodbc==4.0.27
pyparsing==2.4.2
pytest==4.6.5
python-dateutil==2.8.0
PyYAML==5.1.2
pyzmq==18.1.1
qe-client-lib==0.0.1
readline == 6.2.4.1; sys_platform == 'darwin'
requests==2.22.0
s3transfer==0.2.1
sasl==0.2.1
scandir==1.10.0
six==1.13.0
sqlparse==0.1.19
thrift==0.9.3
thrift-sasl==0.3.0
urllib3==1.25.7
wcwidth==0.1.7
Werkzeug==0.16.0
zipp==0.6.0
